// src/payload-types.ts
export interface Media {
  id: number;
  /**
   * 用於 SEO 和無障礙設計
   */
  alt?: string | null;
  caption?: string | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Author {
  id: number;
  name: string;
  slug: string;
  avatar?: (number | null) | Media;
  occupation?: string | null;
  company?: string | null;
  bio?: string | null;
  email?: string | null;
  github?: string | null;
  linkedin?: string | null;
  x?: string | null;
  facebook?: string | null;
  youtube?: string | null;
  instagram?: string | null;
  medium?: string | null;
  updatedAt: string;
  createdAt: string;
}